reorg functional photometry + notebooks
